refactor: replace generateJsonSchemaFiles with makeJsonSchemaFile
