+++ src/utils/index.ts
@@ -1,7 +1,7 @@
 export { patchJsonSchema } from './patchJsonSchema';
 export { clearFolder } from './clearFolder';
 export { jsonSchemaToTsConst } from './jsonSchemaToTsConst';
-export { generateJsonSchemaFiles } from './generateJsonSchemaFiles';
 export { convertOpenApiParameters } from './convertOpenApiParameters';
 export { convertOpenApiToJsonSchema } from './convertOpenApiToJsonSchema';
-export type { JSONSchema, OpenApiSchema } from './types';
+export { makeJsonSchemaFile } from './makeJsonSchemaFile';
+export type { JSONSchema, OpenApiSchema, SchemaPatcher } from './types';
